pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hrm"
version = "0.1.0"
description = "HRM - Human Reasoning Model"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch",
    "adam-atan2",
    "einops",
    "tqdm",
    "coolname",
    "pydantic",
    "argdantic",
    "wandb",
    "omegaconf",
    "hydra-core",
    "huggingface_hub",
    "numpy>=1.24.4",
    "ast2json>=0.4",
    "pyyaml>=6.0.2",
    "mlx>=0.18.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["models*", "utils*", "dataset*", "config*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
